Implemented basic saving, deleting and updating of contact in database using the application

+++ EclipseProject/src/com/akshaykalyan/contact/Contact.java
@@ -264,7 +264,7 @@ public class Contact implements Parcelable {
 	/**
 	 * ID - Getter
 	 */
-	public long getfId() {
+	public int getfId() {
 		return fId;
 	}
